rename quarterly and remove diffy (didn't work)
